Класс информации о запросе на тестирование вынесен в отдельный проект
